+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 FEATURES:
 
+* **New Parameter:** `subnets.enforce_private_link_service_network_policies`
+* **New Parameter:** `subnets.enforce_private_link_endpoint_network_policies`
 * **New Parameter:** `subnets.delegation`: The delegation of the services of `Microsoft.BareMetal/AzureVMware`, `Microsoft.BareMetal/CrayServers`, `Microsoft.Batch/batchAccounts`, `Microsoft.ContainerInstance/containerGroups`, `Microsoft.Databricks/workspaces`, `Microsoft.DBforPostgreSQL/serversv2`, `Microsoft.HardwareSecurityModules/dedicatedHSMs`, `Microsoft.Logic/integrationServiceEnvironments`, `Microsoft.Netapp/volumes`, `Microsoft.ServiceFabricMesh/networks`, `Microsoft.Sql/managedInstances`, `Microsoft.StreamAnalytics/streamingJobs`, `Microsoft.Web/hostingEnvironments` and `Microsoft.Web/serverFarms`.
 
 BUG FIXES: